pyproject.toml
[project]
name = "lm-aig"
version = "0.1.0"
description = "LLM-based Multi-Agent Automatic Item Generation system"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "langchain-openai",
    "langchain-core",
    "pydantic",
    "pydantic-settings",
    "structlog",
    "tavily-python",
    "rich",
    "langchain-groq",
    "langchain-ollama",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-httpx",
    "httpx",
    "langgraph-cli[inmem]",
]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "prometheus-client>=0.21",
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg[binary]>=3.2",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
